Разбивает JS на модули
