+++ index.html
@@ -236,7 +236,7 @@
   </template>
 
   <script src="./js/functions.js"></script>
-  <script src="/js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
   
 </body>
