+++ js/data.js
@@ -1,0 +1,36 @@
+const DESCRIPTIONS = [
+  'Прекрасный закат на побережье',
+  'Уютное кафе в центре города',
+  'Незабываемое путешествие в горы',
+  'Семейный пикник в парке',
+  'Яркие краски осени',
+  'Городской пейзаж в дождливый день',
+  'Веселая вечеринка с друзьями',
+  'Уличное искусство',
+  'Домашние питомцы',
+  'Архитектурные шедевры'
+];
+
+const COMMENTS_MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const NAMES = [
+  'Артём',
+  'Мария',
+  'Иван',
+  'Екатерина',
+  'Александр',
+  'Ольга',
+  'Дмитрий',
+  'Анна',
+  'Сергей',
+  'Наталья'
+];
+
+export {DESCRIPTIONS, COMMENTS_MESSAGES, NAMES};
